Fixed unit test failure
